Separate p5/p5play code into GameLabP5.js
